DA-8 Move prisma module to the config folder
